requirements.txt
# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# LangChain / LangGraph
langchain==0.2.6
langchain-community==0.2.6
langgraph==0.1.13

# Pinecone
pinecone==4.1.0

# Embeddings
sentence-transformers==3.0.1

# Ollama
ollama==0.2.1

# Speech-to-Text
faster-whisper==1.0.3

# Text-to-Speech
gTTS==2.5.1

# Web Scraping & Document Parsing
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2
pypdf==4.2.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.4
numpy==1.26.4
